[Timers] Track timer event processing counter
* Tracking timer events processed at Golang consumer level.
* Include DeleteKV operation under transactions.
* Addresses couple of timer processing bugs:
  - Skip json unmarshal when value is empty
  - Incorrect error matching lead to continuous looping because
    of goto statement

Reviewed-on: http://review.couchbase.org/78124
Tested-by: Abhishek Singh <abhishek@couchbase.com>

+++ consumer/v8_consumer.go
@@ -81,8 +81,8 @@ func (c *Consumer) Serve() {
 	c.stopConsumerCh = make(chan bool, 1)
 	c.stopCheckpointingCh = make(chan bool)
 
-	c.dcpMessagesProcessed = make(map[mcd.CommandCode]int)
-	c.v8WorkerMessagesProcessed = make(map[string]int)
+	c.dcpMessagesProcessed = make(map[mcd.CommandCode]uint64)
+	c.v8WorkerMessagesProcessed = make(map[string]uint64)
 
 	c.consumerSup = suptree.NewSimple(c.workerName)
 	go c.consumerSup.ServeBackground()
